pyproject.toml
[project]
name = "turbodiffusion"
description = "TurboDiffusion: video generation acceleration framework that could accelerate end-to-end video generation by 100-205x with negligible video quality loss."
version = "1.0.0"
authors = [
  { name = "Jintao Zhang" },
  { name = "Kaiwen Zheng" },
  { name = "Kai Jiang" },
  { name = "Haoxu Wang" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
]

dependencies = [
  "setuptools>=62",
  "torch>=2.7.0",
  "torchvision",
  "triton>=3.3.0",
  "flash-attn",
  "einops",
  "numpy",
  "pillow",
  "loguru",
  "imageio[ffmpeg]",
  "pandas",
  "PyYAML",
  "omegaconf",
  "attrs",
  "fvcore",
  "ftfy",
  "regex",
  "transformers",
  "nvidia-ml-py",
  "prompt-toolkit>=3.0",
  "rich>=13.0",
  "peft>=0.11.0",
  "megatron-core",
  "hydra-core",
  "wandb",
  "webdataset",
  "transformer_engine[pytorch]",
  "diffusers",
]

[build-system]
requires = [
  "setuptools>=62",
  "wheel>=0.38",
  "packaging>=21",
  "torch>=2.7.0",
  "ninja",
]
build-backend = "setuptools.build_meta"

[project.scripts]
turbodiffusion-serve = "turbodiffusion.serve.tui:main"

[project.urls]
Homepage = "https://github.com/thu-ml/TurboDiffusion"
Repository = "https://github.com/thu-ml/TurboDiffusion"

# flash-attn 构建时需要 import torch，但未声明 build dependency
[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]